[REM] test_orm: remove `web` dependency
