+++ odoo/addons/test_orm/tests/__init__.py
@@ -17,12 +17,10 @@ from . import (
     test_ir_rules,
     test_json_field,
     test_many2many,
-    test_onchange,
     test_one2many,
     test_performance,
     test_properties,
     test_qweb_float,
-    test_read_group_formatted,
     test_read_group_override,
     test_read_group_private,
     test_read_grouping_sets,
