+++ odoo/addons/test_orm/tests/test_search.py
@@ -1888,7 +1888,5 @@ class TestNonIntId(TransactionCase):
         self.assertEqual(records.name, 'test')
 
     def test_query_non_int_read_group(self):
-        result = self.env['test_orm.view.str.id'].formatted_read_group([], ['name'], ['__count'])
-        self.assertEqual(result, [{'name': 'test', '__count': 1, '__extra_domain': [('name', '=', 'test')]}])
-        result = self.env['test_orm.view.str.id'].formatted_read_group([], [], ['name:count'])
-        self.assertEqual(result, [{'name:count': 1, '__extra_domain': [(1, '=', 1)]}])
+        result = self.env['test_orm.view.str.id']._read_group([], ['name'], ['__count'])
+        self.assertEqual(result, [('test', 1)])
