+++ odoo/addons/test_web/tests/__init__.py
@@ -1,6 +1,12 @@
+from . import test_fields
+from . import test_onchange
+from . import test_properties
+from . import test_read_group_formatted
+from . import test_read_grouping_sets
 from . import test_read_progress_bar
 from . import test_search_panel_select_multi_range
 from . import test_search_panel_select_range
+from . import test_views
 from . import test_web_fill_temporal
 from . import test_web_group_expand
 from . import test_web_read
