+++ odoo/addons/test_web/tests/test_read_group_formatted.py
